+++ src/main/java/com/yulaiz/dong/web/common/interceptor/AuthInterceptor.java
@@ -1,6 +1,9 @@
-package com.yulaiz.dong.web.interceptor;
+package com.yulaiz.dong.web.common.interceptor;
 
 import com.yulaiz.dong.web.common.annotation.IgnoreSecurity;
+import com.yulaiz.dong.web.common.annotation.IgnoreUser;
+import com.yulaiz.dong.web.common.exception.ExeResultException;
+import com.yulaiz.dong.web.common.utils.StringUtil;
 import com.yulaiz.dong.web.model.entity.UserInfo;
 import com.yulaiz.dong.web.service.UserService;
 import lombok.extern.slf4j.Slf4j;
@@ -40,7 +43,7 @@ public class AuthInterceptor extends HandlerInterceptorAdapter {
         Method method = handlerMethod.getMethod();
         String requestPath = request.getRequestURI();
         log.debug("requestIp: " + getIpAddress(request));
-        log.debug("Method: " + method.getName() + ", IgnoreSecurity: " + method.isAnnotationPresent(IgnoreSecurity.class));
+        log.debug("Method: " + method.getName() + ", IgnoreSecurity: " + method.isAnnotationPresent(IgnoreSecurity.class) + ", IgnoreUser: " + method.isAnnotationPresent(IgnoreUser.class));
         log.debug("requestPath: " + requestPath);
         if (requestPath.contains("/v2/api-docs") || requestPath.contains("/swagger") || requestPath.contains("/configuration/ui")) {
             return true;
@@ -53,6 +56,13 @@ public class AuthInterceptor extends HandlerInterceptorAdapter {
         }
         String token = request.getHeader("ACCESS_TOKEN");
         log.debug("token: " + token);
+        if (StringUtil.isEmpty(token)) {
+            throw new ExeResultException("无效token");
+        }
+        request.setAttribute("currentToken", token);
+        if (method.isAnnotationPresent(IgnoreUser.class)) {
+            return true;
+        }
         UserInfo userInfo = userService.getUserByToken(token);
         request.setAttribute("currentUser", userInfo);
         return true;